+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Hide stack trace after solver dependency error
 - Fix packing archives with version from git tag
+- Validate case-insensitive duplicate items in the manifest file
 
 ## [1.1.4] 2022-07-04